土台完成

+++ article/app/Http/Controllers/ArticleController.php
@@ -14,7 +14,7 @@ class ArticleController extends Controller
      */
     public function index()
     {
-        $articles = Article::all();
+        $articles = Article::orderBy('created_at', 'desc')->paginate(10);
         $data = ['articles' => $articles];
         return view('articles.index', $data);
     }
@@ -45,6 +45,7 @@ class ArticleController extends Controller
         ]);
 
         $article = new Article();
+        $article->user_id = \Auth::id();
         $article->title = $request->title;
         $article->body = $request->body;
         $article->save();
@@ -72,6 +73,7 @@ class ArticleController extends Controller
      */
     public function edit(Article $article)
     {
+        $this->authorize($article);
         $data = ['article' => $article];
         return view('articles.edit', $data);
     }
@@ -85,6 +87,7 @@ class ArticleController extends Controller
      */
     public function update(Request $request, Article $article)
     {
+        $this->authorize($article);
         $this->validate($request, [
             'title' => 'required|max:255',
             'body' => 'required'
@@ -104,7 +107,17 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
+        $this->authorize($article);
         $article->delete();
         return redirect(route('articles.index'));
+    }
+
+    public function bookmark_articles() 
+    {
+        $articles = \Auth::user()->bookmark_article()->orderBy('created_at', 'desc')->paginate(10);
+        $data = [
+            'articles' => $articles,
+        ];
+        return view('articles.bookmarks', $data);
     }
 }
